Segregando em fila a importação do acervo tridimensionais
