+++ SME.CDEP.Aplicacao/Servicos/ServicoImportacaoArquivoAcervoTridimensional.cs
@@ -8,6 +8,7 @@ using SME.CDEP.Dominio.Constantes;
 using SME.CDEP.Dominio.Entidades;
 using SME.CDEP.Dominio.Excecoes;
 using SME.CDEP.Dominio.Extensions;
+using SME.CDEP.Infra;
 using SME.CDEP.Infra.Dados.Repositorios.Interfaces;
 using SME.CDEP.Infra.Dominio.Enumerados;
 
@@ -15,22 +16,17 @@ namespace SME.CDEP.Aplicacao.Servicos
 {
     public class ServicoImportacaoArquivoAcervoTridimensional : ServicoImportacaoArquivoBase, IServicoImportacaoArquivoAcervoTridimensional 
     {
-        private readonly IServicoAcervoTridimensional servicoAcervoTridimensional;
+        private readonly IServicoMensageria servicoMensageria;
         
         public ServicoImportacaoArquivoAcervoTridimensional(IRepositorioImportacaoArquivo repositorioImportacaoArquivo, IServicoMaterial servicoMaterial,
             IServicoEditora servicoEditora,IServicoSerieColecao servicoSerieColecao,IServicoIdioma servicoIdioma, IServicoAssunto servicoAssunto,
             IServicoCreditoAutor servicoCreditoAutor,IServicoConservacao servicoConservacao, IServicoAcessoDocumento servicoAcessoDocumento,
-            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte, IServicoFormato servicoFormato,IServicoAcervoTridimensional servicoAcervoTridimensional, 
+            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte, IServicoFormato servicoFormato,IServicoMensageria servicoMensageria, 
             IMapper mapper,IRepositorioParametroSistema repositorioParametroSistema)
             : base(repositorioImportacaoArquivo, servicoMaterial, servicoEditora,servicoSerieColecao, servicoIdioma, servicoAssunto, servicoCreditoAutor,
                 servicoConservacao,servicoAcessoDocumento,servicoCromia,servicoSuporte,servicoFormato, mapper,repositorioParametroSistema)
         {
-            this.servicoAcervoTridimensional = servicoAcervoTridimensional ?? throw new ArgumentNullException(nameof(servicoAcervoTridimensional));
-        }
-
-        public void DefinirCreditosAutores(List<IdNomeTipoDTO> creditosAutores)
-        {
-            CreditosAutores = creditosAutores;
+            this.servicoMensageria = servicoMensageria ?? throw new ArgumentNullException(nameof(servicoMensageria));
         }
         
         public async Task<bool> RemoverLinhaDoArquivo(long id, LinhaDTO linhaDoArquivo)
@@ -85,18 +81,18 @@ namespace SME.CDEP.Aplicacao.Servicos
         
             var importacaoArquivo = ObterImportacaoArquivoParaSalvar(file.FileName, TipoAcervo.Tridimensional, JsonConvert.SerializeObject(acervosTridimensionalLinhas));
             
-            await CarregarTodosOsDominios();
-            
-            var importacaoArquivoId = await PersistirImportacao(importacaoArquivo);
+           var importacaoArquivoId = await PersistirImportacao(importacaoArquivo);
            
-            ValidarPreenchimentoValorFormatoQtdeCaracteres(acervosTridimensionalLinhas);
-            
-            await PersistenciaAcervo(acervosTridimensionalLinhas);
-            await AtualizarImportacao(importacaoArquivoId, JsonConvert.SerializeObject(acervosTridimensionalLinhas), acervosTridimensionalLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
-        
-            var arquivoImportado = await repositorioImportacaoArquivo.ObterPorId(importacaoArquivoId);
-        
-            return await ObterRetornoImportacaoAcervo(arquivoImportado, acervosTridimensionalLinhas);
+           await servicoMensageria.Publicar(RotasRabbit.ExecutarImportacaoArquivoAcervoTridimensional, importacaoArquivoId, Guid.NewGuid());
+
+            return new ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoTridimensionalDTO,AcervoTridimensionalLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>()
+            {
+                Id = importacaoArquivoId,
+                Nome = importacaoArquivo.Nome,
+                TipoAcervo = importacaoArquivo.TipoAcervo,
+                DataImportacao = DateTimeExtension.HorarioBrasilia(),
+                Status = ImportacaoStatus.Pendente
+            };
         }
         
         private async Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoTridimensionalDTO,AcervoTridimensionalLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ObterRetornoImportacaoAcervo(ImportacaoArquivo arquivoImportado, IEnumerable<AcervoTridimensionalLinhaDTO> acervosTridimensionalLinhas, bool estaImportandoArquivo = true)
@@ -211,85 +207,6 @@ namespace SME.CDEP.Aplicacao.Servicos
 		        mensagemErro.Add(acervoTridimensionalLinhaDTO.Diametro.Mensagem);
 			        
             return mensagemErro.ToArray();
-        }
-        
-        public async Task PersistenciaAcervo(IEnumerable<AcervoTridimensionalLinhaDTO> acervosTridimensionalsLinhas)
-        {
-            foreach (var acervoTridimensionalLinha in acervosTridimensionalsLinhas.Where(w=> !w.PossuiErros))
-            {
-                try
-                {
-                    var acervoTridimensional = new AcervoTridimensionalCadastroDTO()
-                    {
-                        Titulo = acervoTridimensionalLinha.Titulo.Conteudo,
-                        Codigo = acervoTridimensionalLinha.Codigo.Conteudo,
-                        Procedencia = acervoTridimensionalLinha.Procedencia.Conteudo,
-                        Ano = acervoTridimensionalLinha.Ano.Conteudo,
-                        ConservacaoId = ObterConservacaoIdPorValorDoCampo(acervoTridimensionalLinha.EstadoConservacao.Conteudo),
-                        Quantidade = acervoTridimensionalLinha.Quantidade.Conteudo.ConverterParaInteiro(),
-                        Descricao = acervoTridimensionalLinha.Descricao.Conteudo,
-                        Largura = acervoTridimensionalLinha.Largura.Conteudo,
-                        Altura = acervoTridimensionalLinha.Altura.Conteudo,
-                        Profundidade = acervoTridimensionalLinha.Profundidade.Conteudo,
-                        Diametro = acervoTridimensionalLinha.Diametro.Conteudo,
-                    };
-                    await servicoAcervoTridimensional.Inserir(acervoTridimensional);
-        
-                    acervoTridimensionalLinha.DefinirLinhaComoSucesso();
-                }
-                catch (Exception ex)
-                {
-                    acervoTridimensionalLinha.DefinirLinhaComoErro(ex.Message);
-                }
-            }
-        }
-
-        public void ValidarPreenchimentoValorFormatoQtdeCaracteres(IEnumerable<AcervoTridimensionalLinhaDTO> linhas)
-        {
-            foreach (var linha in linhas)
-            {
-                try
-                {
-                    ValidarPreenchimentoLimiteCaracteres(linha.Titulo, Constantes.TITULO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Codigo, Constantes.TOMBO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Procedencia,Constantes.PROCEDENCIA);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Ano,Constantes.ANO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.EstadoConservacao,Constantes.ESTADO_CONSERVACAO);
-                    ValidarConteudoCampoComDominio(linha.EstadoConservacao, Conservacoes, Constantes.ESTADO_CONSERVACAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Quantidade,Constantes.QUANTIDADE);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Descricao,Constantes.DESCRICAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Largura,Constantes.LARGURA);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Altura,Constantes.ALTURA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Profundidade,Constantes.PROFUNDIDADE);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Diametro,Constantes.DIAMETRO);
-                    
-                    linha.PossuiErros = PossuiErro(linha);
-                }
-                catch (Exception e)
-                {
-                    linha.DefinirLinhaComoErro(string.Format(Constantes.OCORREU_UMA_FALHA_INESPERADA_NA_LINHA_X_MOTIVO_Y, linha.NumeroLinha, e.Message));
-                }
-            }
-        }
-
-        private bool PossuiErro(AcervoTridimensionalLinhaDTO linha)
-        {
-            return linha.Titulo.PossuiErro 
-                   || linha.Codigo.PossuiErro 
-                   || linha.Procedencia.PossuiErro
-                   || linha.Ano.PossuiErro
-                   || linha.EstadoConservacao.PossuiErro
-                   || linha.Quantidade.PossuiErro
-                   || linha.Descricao.PossuiErro
-                   || linha.Largura.PossuiErro
-                   || linha.Altura.PossuiErro
-                   || linha.Profundidade.PossuiErro
-                   || linha.Diametro.PossuiErro;
         }
         
         private async Task<IEnumerable<AcervoTridimensionalLinhaDTO>> LerPlanilha(IFormFile file)
