+++ SME.CDEP.Aplicacao/UseCase/Interface/IExecutarImportacaoArquivoAcervoTridimensionalUseCase.cs
@@ -1,0 +1,8 @@
+
+using SME.CDEP.Infra;
+
+namespace SME.CDEP.Aplicacao
+{
+    public interface IExecutarImportacaoArquivoAcervoTridimensionalUseCase : IRabbitUseCase
+    {}
+}
