+++ SME.CDEP.TesteIntegracao/ImportacaoArquivo/Ao_fazer_manutencao_importacao_arquivo_acervo_tridimensional.cs
@@ -32,8 +32,8 @@ namespace SME.CDEP.TesteIntegracao
             acervoTridimensionalLinhas[8].Codigo.Conteudo = string.Empty;
             var linhasComErros = new[] { 3, 5, 8, 9 };
             
-            await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
 
             foreach (var linha in acervoTridimensionalLinhas)
             {
@@ -109,8 +109,8 @@ namespace SME.CDEP.TesteIntegracao
             acervoTridimensionalLinhas[8].EstadoConservacao.Conteudo = "Desconhecido";
             var linhasComErros = new[] { 3, 5, 8, 9 };
             
-            await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
 
             foreach (var linha in acervoTridimensionalLinhas)
             {
@@ -193,8 +193,8 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoTridimensionalLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoTridimensionalLinhas);
         
             var acervos = ObterTodos<Acervo>();
             var acervosTridimensional = ObterTodos<AcervoTridimensional>();
@@ -256,9 +256,9 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoTridimensionalLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoTridimensionalLinhas);
             await servicoImportacaoArquivo.AtualizarImportacao(1, JsonConvert.SerializeObject(acervoTridimensionalLinhas), acervoTridimensionalLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
             var retorno = await servicoImportacaoArquivo.ObterImportacaoPendente();
 
@@ -368,9 +368,9 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoTridimensionalLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosTridimensionais();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoTridimensionalLinhas);
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoTridimensionalLinhas);
             await servicoImportacaoArquivo.AtualizarImportacao(1, JsonConvert.SerializeObject(acervoTridimensionalLinhas), acervoTridimensionalLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
             var retorno = await servicoImportacaoArquivo.ObterImportacaoPendente();
