+++ SME.CDEP.Worker/WorkerRabbitMQ.cs
@@ -47,6 +47,7 @@ namespace SME.CDEP.Worker
             Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoArteGrafica, new ComandoRabbit("Realiza a importação de acervos artes gráficas por planilha", typeof(IExecutarImportacaoArquivoAcervoArteGraficaUseCase), true));
             Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoAudiovisual, new ComandoRabbit("Realiza a importação de acervos audiovisual por planilha", typeof(IExecutarImportacaoArquivoAcervoAudiovisualUseCase), true));
             Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoFotografico, new ComandoRabbit("Realiza a importação de acervos fotográficos por planilha", typeof(IExecutarImportacaoArquivoAcervoFotograficoUseCase), true));
+            Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoTridimensional, new ComandoRabbit("Realiza a importação de acervos tridimensionais por planilha", typeof(IExecutarImportacaoArquivoAcervoTridimensionalUseCase), true));
         }
     }
 }
